mod:: 데이터 전처리 함수 lectureUtil.ts로 분리 - dayMap, dataPreprocess를 lectureUtil.ts 파일로 분리
